Implemented heuristic to speed up the test
this heuristic verifies if the last combination of joint 0 + joint 1 + all possible positions of joint 2 resulted in a collision
If no collision was detected, it is because joint 1 is high enough that the arm is far enough from the torso. Any other position
for joint 1 from here on would never result in a collision, so we can skip them and increment joint 0 position instead.